requirements.txt
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///tmp/build/80754af9/beautifulsoup4_1631874778482/work
black==22.8.0
brotlipy==0.7.0
certifi==2021.10.8
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
chardet @ file:///tmp/build/80754af9/chardet_1607706746162/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
commonmark==0.9.1
conda==4.11.0
conda-build==3.21.7
conda-content-trust @ file:///tmp/build/80754af9/conda-content-trust_1617045594566/work
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1618262148928/work
cryptography @ file:///tmp/build/80754af9/cryptography_1639400846433/work
cycler==0.11.0
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
dnspython==2.2.1
filelock @ file:///opt/conda/conda-bld/filelock_1642510437405/work
fonttools==4.37.1
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
ipython @ file:///tmp/build/80754af9/ipython_1643818147236/work
jedi @ file:///tmp/build/80754af9/jedi_1644315233700/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
kiwisolver==1.4.4
libarchive-c @ file:///tmp/build/80754af9/python-libarchive-c_1617780486945/work
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1621528148836/work
matplotlib==3.5.3
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mypy-extensions==0.4.3
nibabel==4.0.2
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1634095647912/work
packaging==21.3
pandas==1.4.4
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathspec==0.10.1
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.0.1
pkginfo @ file:///tmp/build/80754af9/pkginfo_1643162084911/work
platformdirs==2.5.2
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
psutil @ file:///tmp/build/80754af9/psutil_1612298023621/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1635333100036/work
pyparsing==3.0.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-etcd==0.4.5
pytz==2021.3
PyYAML==6.0
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
rich==12.5.1
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016699510/work
scipy==1.9.1
seaborn==0.11.2
six @ file:///tmp/build/80754af9/six_1623709665295/work
soupsieve @ file:///tmp/build/80754af9/soupsieve_1636706018808/work
tomli==2.0.1
torch==1.11.0
torchelastic==0.2.2
torchtext==0.12.0
torchvision==0.12.0
tqdm @ file:///tmp/build/80754af9/tqdm_1635330843403/work
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing_extensions==4.3.0
urllib3==1.26.7
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
